Move builtin loading code entirely out of __init__
... YAY!

This paves the way for further improvements.
