+++ mathics/core/definitions.py
@@ -15,9 +15,10 @@ from mathics.core.attributes import A_NO_ATTRIBUTES
 from mathics.core.convert.expression import to_mathics_list
 from mathics.core.element import fully_qualified_symbol_name
 from mathics.core.expression import Expression
-from mathics.core.load_builtin import definition_contribute
+from mathics.core.load_builtin import definition_contribute, modules
 from mathics.core.symbols import Atom, Symbol, strip_context
 from mathics.core.systemsymbols import SymbolGet
+from mathics.settings import ROOT_DIR
 
 type_compiled_pattern = type(re.compile("a.a"))
 
@@ -138,8 +139,6 @@ class Definitions:
         self.timing_trace_evaluation = False
 
         if add_builtin:
-            from mathics.builtin import modules
-            from mathics.settings import ROOT_DIR
 
             loaded = False
             if builtin_filename is not None:
