+++ mathics/version.py
@@ -5,4 +5,4 @@
 # well as importing into Python. That's why there is no
 # space around "=" below.
 # fmt: off
-__version__="6.0.2dev0"  # noqa
+__version__="6.1.0dev0"  # noqa
